Refactor moving elements
Now the moveHandler only uses the mouse-startpoint
and  end-point to calculate the difference (delta) between them.
Actual layouting the components is done by the component itself.

The components calculate their position by: givenDelta + this.getXY
givenDelta is the delta given from the moveHandler and getXY
is the old position of the shape.

This is the behaviour you would expect from a move-function/handler!