Fix abort draw-process if someone changes the tool
When someone changes the drawing-tool (e.g.: to the arrow) the
drawing-process (if any) should be aborted. For example: if someone
draws a path and then chooses the rectangle-tool the drawing-process of
the path stops, kills all previous temporary-drawn-shapes and starts
with the rectangle tool.

Without this process you get really strange behaviour when switching
betwen pathes and polygons. ;-)