Fix bug: moving img while resizing
The image was moved along the path with which the img should get
resized. That came because the image preserve's it's ratio so the image
isn't always as big as the full path.

This fix calculates the new x/y-coordinate by the new set width/height
of the image/RectangleLike.