Move config-files into ~/.move
Config files are now stored in a move-config directory inside user's
home directory. The original approach (place them on the classpath)
doesn't allow altering the file and places the configfiles inside the
published jar.

The files are still published through the jar and classpath but get
copied into the ~/.move directory if they don't exist.