Create PathLike for abstract over path-like elements
This trait represents shapes which are represented by an
underlying path. It's similar to RectangleLike where each
implementing class is a shape that looks like an rectangle.

Implementing classes of PathLike only need to give 2 implementations:
- resize() for resizing a specific edge-point identified by an index
- getEdgePoint() for getting the point at a given index
The implementing classes get resizing, moving for free!
(They will get rotation for free in the future!)