Fix buggy resizing of rotated rectangles
When resizing rotated rectangles you need to carefully calculate the
new x's,y's, width's & height's.
Main problem is that the middle-point of the rectangle, which is also
the pivot point of the rotation, is changing through the resize the
problem get solved through GeometryUtils.calculateRotationOffset().

Another major problem are the coordinates of the mouse events.
Mouse-event's source are the anchors, therefore the coordinates of the mouse are
RELATIVE TO THE ANCHOR and not the rectangle. You need to translate the
mouse-event to an event relative to the scene and translate these
coordinates into the coordinate-system of the rectangle using
sceneToLocal(). This is done by calculateDelta().
Affter that you can use the function
GeometryUtils.calculateRotationOffset() to calculate an offset which
gets added to the X/Y-coordinates as additional offset.