Fix resizing circles
When resizing a circle the circle resized in false directions. E.g:
If you expected that the circle get smaller it got bigger. If you
expected the circle get bigger it got sporadically smaller.

Fixed this by dividing the parameters of setWidth,setHeight by 2 and
doubling the radius as return-value for getWidth,getHeight.